+++ organizer/classifiers/dnn_classifier.py
@@ -58,7 +58,7 @@ class DnnClassifier(Classifier):
 
     def get_classification(
         self,
-        evaluation_dict: dict[str, int],
+        evaluation_dict: dict[str, float],
         tags: Iterable[str],
         print_buffer: PrintBuffer,
     ) -> str:
