+++ organizer/classifiers/random_forest_classifier.py
@@ -36,7 +36,7 @@ class RandomForestClassifier(Classifier):
 
     def get_classification(
         self,
-        evaluation_dict: dict[str, int],
+        evaluation_dict: dict[str, float],
         tags: Iterable[str],
         print_buffer: PrintBuffer,
     ) -> str:
